LivingEntity class created
