+++ Assets/AbstractClasses.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1b27f94724d617646ba31582f0e47296
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
