+++ Assets/AbstractClasses/LivingEntity.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 10bfa9291c45d1c449edcd16604e7dce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
