+++ Assets/tmp.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ba2757bbaf46ce648b1caacdba3967b2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
